Rename a mis-named "attribute-db" into a "zcl-data-model", since that's what it really is.
