+++ src/app/plugin/on-off-server/Makefile
@@ -1,5 +1,5 @@
 NAME=on-off-server
-REQUIRED_PLUGINS=../attribute-db/attribute-db.o ../../utest-gen/gen-callback-stubs.o
+REQUIRED_PLUGINS=../zcl-data-model/zcl-data-model.o ../../utest-gen/gen-callback-stubs.o
 
 # ---------- GENERIC -------------
 CC=gcc
